Changes to Adding/Removing Actors
Added in a new way of adding and removing actors changing the storage from a Dictionary to a list of the new class CharacterInScene.  
CutsceneEditor now creates a temporary game object in the scene and stores the characters in the scene under it so they can be accessed by the editor properly.