Cutscene Text Saving
Cutscenes can now be saved as text.
Added in buttons to the Cutscene Editor UI to allow for easy saving/loading of text file representation.
Fixed an issue with the Dialog Editor UI.
Fixed an issue where the Cutscene Editor UI would not appear or disappear leaving an extra instance of the Cutscene gameobject.