Cutscene Editor Issues Fixed
Fixed an error where the steps in the cutscene editor were not cleared when loading a cutscene from a text file.
Dialog is now always the last element of a cutscene step to prevent a dangling "and" at the end for the next element.  There is also now a one dialog per step restriction.
Fixed an issue where the MovementEditor was not properly parsing when followed by "and".