More Editor Additions
Cutscene Text files will now have a line at the top to note which scene it supposed to take place in.
Fixed an issue where CutsceneDialogs weren't being loaded properly for dialog with no listed speaker.
The name of the text file storing a cutscene is now used as the name of the cutscene.
Added in a new JoinCamelCase method to extension methods.