Cutscene Changes
Moved all of the parsing that goes into cutscenes into a dedicated CutsceneParser class.
Changed how sequences of CutsceneElements are stored to use a CutsceneStep to keep in line with the CutsceneEditor.  What had been previously been a CutsceneStep in the Editor is now a CutsceneStepEditor.
Fixed an issue that was preventing the final pan from the opening cutscene from finishing properly.